+++ lambdas/update_supporter_appointment_settings.py
@@ -1,6 +1,7 @@
 #Written by Nhan Le
 
 from package.query_db import query
+from package.dictionary_to_list import dictionary_to_list
 import json
 
 def update_supporter_appointment_settings(event, context):
@@ -65,19 +66,3 @@ def update_supporter_appointment_settings(event, context):
     return {
         'statusCode': 200
     }
-
-def dictionary_to_list(dictionary):
-    params = []
-    valueType = ""
-    #For each key, name = key and value = dictionary {value type : value}
-    for key in dictionary.keys():
-        if(type(dictionary[key]) == int):
-            valueType = 'longValue'
-        if(type(dictionary[key]) == str):
-            valueType = 'stringValue'
-        if(type(dictionary[key]) == bool):
-            valueType = 'booleanValue'
-        if(type(dictionary[key]) == float):
-            valueType = 'doubleValue'
-        params.append({'name': key, 'value': {valueType : dictionary[key]}})
-    return params
